[X11RootWindow] Fix image wallpaper not being set properly
Picom (this probably applies to other compositors) couldn't fetch the wallpaper because it was destroyed at the end of the ImageWallpaper instance's life. Adding "XSetCloseDownMode(_display, RetainPermanent);" in the constructor (this can be moved so it is done only once) and removing the call to XFreePixmap in the destructor fixed the problem.